OK-513: Refaktoroitu "valintalaskennan-tulos"-sivua selkeämmäksi
Siirretty koodia eri tiedostoihin ja jaettu koodia pienempiin komponentteihin
