+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valintalaskennan-tulos/page.tsx
@@ -1,247 +1,22 @@
 'use client';
 
 import { TabContainer } from '../tab-container';
-import { TablePaginationWrapper } from '@/app/components/table/table-pagination-wrapper';
 import { QuerySuspenseBoundary } from '@/app/components/query-suspense-boundary';
-import {
-  Button,
-  Box,
-  Accordion,
-  AccordionSummary,
-  AccordionDetails,
-} from '@mui/material';
+import { Button, Box } from '@mui/material';
 import { useTranslations } from '@/app/hooks/useTranslations';
-import { LaskettuValintatapajonoTable } from './laskettu-valintatapajono-table';
 import { useLasketutValinnanVaiheet } from '@/app/hooks/useLasketutValinnanVaiheet';
-import {
-  useJonosijatSearch,
-  useJonosijatSearchParams,
-} from '@/app/hooks/useJonosijatSearch';
-import {
-  JonoSijaWithHakijaInfo,
-  LaskettuValinnanVaihe,
-  LaskettuValintatapajono,
-  muutaSijoittelunStatus,
-} from '@/app/lib/valintalaskenta-service';
-import { DEFAULT_BOX_BORDER, OPH_ORGANIZATION_OID } from '@/app/lib/constants';
-import { Typography } from '@opetushallitus/oph-design-system';
 import { ValintalaskennanTulosSearch } from './valintalaskennan-tulos-search';
 import { PageSizeSelector } from '@/app/components/table/page-size-selector';
-import { ExpandMore, FileDownloadOutlined } from '@mui/icons-material';
+import { FileDownloadOutlined } from '@mui/icons-material';
 import { configuration } from '@/app/lib/configuration';
 import React from 'react';
-import { toFormattedDateTimeString } from '@/app/lib/localization/translation-utils';
-import { useUserPermissions } from '@/app/hooks/useUserPermissions';
-import { useHakukohde } from '@/app/hooks/useHakukohde';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Spinner } from '@/app/components/spinner';
+import { ValintatapajonoContent } from './valintatapajono-content';
+import { useJonosijatSearchParams } from '@/app/hooks/useJonosijatSearch';
 
 type LasketutValinnanvaiheetParams = {
   hakuOid: string;
   hakukohdeOid: string;
-};
-
-const ValintatapaJonoTable = ({
-  jonoId,
-  jonosijat,
-}: {
-  jonoId: string;
-  jonosijat: Array<JonoSijaWithHakijaInfo>;
-}) => {
-  const { results, pageResults, sort, setSort, pageSize, setPage, page } =
-    useJonosijatSearch(jonoId, jonosijat);
-
-  return (
-    <TablePaginationWrapper
-      totalCount={results?.length ?? 0}
-      countTranslationKey="jonosijat.maara"
-      pageSize={pageSize}
-      setPageNumber={setPage}
-      pageNumber={page}
-      countHidden={true}
-    >
-      <LaskettuValintatapajonoTable
-        setSort={setSort}
-        sort={sort}
-        jonoId={jonoId}
-        jonosijat={pageResults}
-      />
-    </TablePaginationWrapper>
-  );
-};
-
-const AccordionBlock = ({
-  title,
-  children,
-}: {
-  title: React.ReactNode;
-  children: React.ReactNode;
-}) => {
-  const headerId = `${title}-accordion-header`;
-  const contentId = `${title}-accordion-content`;
-
-  return (
-    <Accordion
-      defaultExpanded={true}
-      sx={{
-        border: DEFAULT_BOX_BORDER,
-      }}
-    >
-      <AccordionSummary
-        expandIcon={<ExpandMore />}
-        aria-controls={contentId}
-        id={headerId}
-      >
-        {title}
-      </AccordionSummary>
-      <AccordionDetails
-        sx={{
-          borderTop: DEFAULT_BOX_BORDER,
-        }}
-      >
-        {children}
-      </AccordionDetails>
-    </Accordion>
-  );
-};
-
-const getJonoNimi = ({
-  valinnanVaiheNimi,
-  jonoNimi,
-}: {
-  valinnanVaiheNimi: string;
-  jonoNimi: string;
-}) => {
-  return jonoNimi.includes(valinnanVaiheNimi)
-    ? jonoNimi
-    : `${valinnanVaiheNimi}: ${jonoNimi}`;
-};
-
-const ValintatapajonoContent = ({
-  hakukohdeOid,
-  valinnanVaihe,
-  jono,
-}: {
-  hakukohdeOid: string;
-  valinnanVaihe: LaskettuValinnanVaihe;
-  jono: LaskettuValintatapajono & { jonosijat: Array<JonoSijaWithHakijaInfo> };
-}) => {
-  const { t } = useTranslations();
-  const jonoSubHeader = `(${toFormattedDateTimeString(valinnanVaihe.createdAt)} | 
-  ${t('yleinen.prioriteetti')}: ${jono.prioriteetti})`;
-
-  const { data: permissions } = useUserPermissions();
-
-  const { data: hakukohde } = useHakukohde({ hakukohdeOid });
-
-  const hasOphUpdate =
-    permissions.writeOrganizations.includes(OPH_ORGANIZATION_OID);
-
-  const hasOrgCrud = permissions.crudOrganizations.includes(
-    hakukohde.organisaatioOid,
-  );
-  const queryClient = useQueryClient();
-
-  const sijoittelunStatusMutation = useMutation({
-    mutationFn: async ({
-      jono,
-      status,
-    }: {
-      jono: LaskettuValintatapajono;
-      status: boolean;
-    }) => {
-      await muutaSijoittelunStatus({ jono, status });
-      queryClient.setQueryData(
-        ['getLasketutValinnanVaiheet', hakukohdeOid],
-        (vaiheet: Array<LaskettuValinnanVaihe>) =>
-          vaiheet.map((vaihe) => ({
-            ...vaihe,
-            valintatapajonot: vaihe.valintatapajonot?.map((oldJono) => ({
-              ...oldJono,
-              valmisSijoiteltavaksi:
-                jono.oid === oldJono.oid ? status : jono.valmisSijoiteltavaksi,
-            })),
-          })),
-      );
-    },
-    onError: (e) => {
-      // TODO: Toast-notifikaatio
-      window.alert('Tapahtui virhe!');
-      console.error(e);
-    },
-  });
-
-  return (
-    <Box key={jono.oid} width="100%">
-      <AccordionBlock
-        title={
-          <Typography
-            variant="h2"
-            component="h3"
-            sx={{
-              display: 'flex',
-              flexDirection: 'row',
-              flexWrap: 'nowrap',
-              columnGap: 2,
-              alignItems: 'center',
-            }}
-          >
-            <div>
-              {getJonoNimi({
-                valinnanVaiheNimi: valinnanVaihe.nimi,
-                jonoNimi: jono.nimi,
-              })}
-            </div>
-            <Typography component="div" variant="body1">
-              {jonoSubHeader}
-            </Typography>
-          </Typography>
-        }
-      >
-        <Box paddingTop={2} paddingBottom={1}>
-          {!jono.valmisSijoiteltavaksi && jono.siirretaanSijoitteluun && (
-            <Button
-              startIcon={
-                sijoittelunStatusMutation.isPending && (
-                  <Spinner color="inherit" size="24px" />
-                )
-              }
-              disabled={
-                sijoittelunStatusMutation.isPending ||
-                (!hasOphUpdate && !hasOrgCrud)
-              }
-              variant="outlined"
-              onClick={() =>
-                sijoittelunStatusMutation.mutate({ jono, status: true })
-              }
-            >
-              {t('valintalaskennan-tulos.siirra-jono-sijoitteluun')}
-            </Button>
-          )}
-          {jono.valmisSijoiteltavaksi && jono.siirretaanSijoitteluun && (
-            <Button
-              startIcon={
-                sijoittelunStatusMutation.isPending && (
-                  <Spinner color="inherit" size="24px" />
-                )
-              }
-              disabled={sijoittelunStatusMutation.isPending || !hasOphUpdate}
-              variant="outlined"
-              onClick={() =>
-                sijoittelunStatusMutation.mutate({ jono, status: false })
-              }
-            >
-              {t('valintalaskennan-tulos.poista-jono-sijoittelusta')}
-            </Button>
-          )}
-        </Box>
-        <ValintatapaJonoTable
-          jonoId={jono.valintatapajonooid}
-          jonosijat={jono.jonosijat}
-        />
-      </AccordionBlock>
-    </Box>
-  );
 };
 
 const LasketutValinnanVaiheetContent = ({
